Rename to network config and update README
