+++ example_network_config.yaml
