+++ util/config/__init__.py
@@ -1,0 +1,2 @@
+from util.config.dataset_config import DatasetConfig, load_data_config
+from util.config.network_config import NetworkConfig, load_network_config
